refactor to MVC
