+++ src/ca/sheridancollege/project/model/Card.java
@@ -1,0 +1,28 @@
+/**
+ * SYST 17796 Project Base code.
+ * Students can modify and extend to implement their game.
+ * Add your name as an author and the date!
+ */
+package ca.sheridancollege.project.model;
+
+/**
+ * A class to be used as the base Card class for the project. Must be general
+ * enough to be instantiated for any Card
+ * game. Students wishing to add to the code should remember to add themselves
+ * as a modifier.
+ *
+ * @author dancye
+ */
+public abstract class Card {
+    // default modifier for child classes
+
+    /**
+     * Students should implement this method for their specific children classes
+     *
+     * @return a String representation of a card. Could be an UNO card, a regular
+     *         playing card etc.
+     */
+    @Override
+    public abstract String toString();
+
+}
