+++ src/ca/sheridancollege/project/model/Game.java
@@ -3,12 +3,13 @@
  * Students can modify and extend to implement their game.
  * Add your name as an author and the date!
  */
-package ca.sheridancollege.project;
+package ca.sheridancollege.project.model;
 
 import java.util.ArrayList;
 
 /**
- * The class that models your game. You should create a more specific child of this class and instantiate the methods
+ * The class that models your game. You should create a more specific child of
+ * this class and instantiate the methods
  * given.
  *
  * @author dancye
@@ -16,7 +17,7 @@ import java.util.ArrayList;
  */
 public abstract class Game {
 
-    private final String name;//the title of the game
+    private final String name;// the title of the game
     private ArrayList<Player> players;// the players of the game
 
     public Game(String name) {
@@ -46,13 +47,15 @@ public abstract class Game {
     }
 
     /**
-     * Play the game. This might be one method or many method calls depending on your game.
+     * Play the game. This might be one method or many method calls depending on
+     * your game.
      */
     public abstract void play();
 
     /**
-     * When the game is over, use this method to declare and display a winning player.
+     * When the game is over, use this method to declare and display a winning
+     * player.
      */
     public abstract void declareWinner();
 
-}//end class
+}// end class
