+++ src/ca/sheridancollege/project/model/Player.java
@@ -3,17 +3,18 @@
  * Students can modify and extend to implement their game.
  * Add your name as an author and the date!
  */
-package ca.sheridancollege.project;
+package ca.sheridancollege.project.model;
 
 /**
- * A class that models each Player in the game. Players have an identifier, which should be unique.
+ * A class that models each Player in the game. Players have an identifier,
+ * which should be unique.
  *
  * @author dancye
  * @author Paul Bonenfant Jan 2020
  */
 public abstract class Player {
 
-    private String name; //the unique name for this player
+    private String name; // the unique name for this player
 
     /**
      * A constructor that allows you to set the player's unique ID
@@ -41,7 +42,8 @@ public abstract class Player {
     }
 
     /**
-     * The method to be overridden when you subclass the Player class with your specific type of Player and filled in
+     * The method to be overridden when you subclass the Player class with your
+     * specific type of Player and filled in
      * with logic to play your game.
      */
     public abstract void play();
