+++ src/ca/sheridancollege/project/model/PokerHandRank.java
@@ -1,4 +1,4 @@
-package ca.sheridancollege.project;
+package ca.sheridancollege.project.model;
 
 public enum PokerHandRank {
     HIGH_CARD("High Card"),
