+++ view.md
@@ -1,0 +1,16 @@
+1. Display welcome screen
+2. Ask how many players are playing the game 3. For each player, ask for their name
+3. Display the game board with player name and their chip count
+   1. table with 3 columns
+      1. player name
+      2. chip count
+      3. bet amount
+4. Show the first player their cards and ask them to make their move
+   1. check
+   2. bet
+   3. raise
+   4. fold
+5. continue asking all players to make their move until bet amount is equal
+6. show the pot amount and evaluate the winner for the round.
+7. eleminate the player id they have no chips left and display elemination after the round.
+8. continue until only one player is left with chips
